// ReportingDashboard/client/src/app/shared/models/theme.ts
export interface Theme {
  windowTitle: string;
  lightColor: string;
  linkColor: string;
  accentColor: string;
  secondaryColor: string;
  grayColor: string;
  lightGrayColor: string;
  lightestGrayColor: string;
  mediumGrayColor: string;
  darkGrayColor: string;
  darkestGrayColor: string;
  logo: string;
  logoPublicId: string;
  logoName: string;
  logoNamePublicId: string;
  favIcon: string;
  favIconPublicId: string;
}